+++ test/test_pekwm.cc
@@ -79,7 +79,6 @@ main_tests(int argc, char *argv[])
 	TestWorkspaces testWorkspaces;
 
 	// x11
-	TestGeometry testGeometry;
 	TestX11 testX11;
 
 	return TestSuite::main(argc, argv);